refactor: move InjuryStatus class to own package
